+++ doc/source/changelog.rst
@@ -47,6 +47,8 @@ This is release version of ``0.9.0`` of Nuclei SDK, which is still under develop
 
   - Only enable/disable L2 cache enable bit during pre-main initialization
   - Limit sPMP/sMPU entry numbers to 16 event PMP entries are 64 for evalsoc to match Nuclei CPU design
+  - Optimize SMP bringup process, boot hart should wait for other harts software pending bit be set, mainly to be suitable
+    for Nuclei Qemu 2025.10, the other harts run slow than main harts
 
 * Documentation